+++ NEWS.md
@@ -1,6 +1,18 @@
 # version 0.9-8
 
-* allow to `st_cast` COMPOUNDCURVE, MULTISURFACE or CURVEPOLYGON to GEOMETRYCOLLECTION (and back); 
+* `distinct.sf` works by default on all variables, and keeps active geometry active; #1613
+
+* improve (fix?) polygonize/contour code; #1608
+
+* `sf_proj_network()` reports whether PROJ uses network (CDN) grids, can switch it on or off, and can set the CDN url.
+
+* `st_write` returns obj, invisibly; #1597
+
+* fix regression in n-ary `st_intersection()`, #1595, introduced at #1549
+
+* `st_inscribed_circle` computes the maximum inscribed circle for polygons (requires GEOS >= 3.9.0)
+
+* allow to `st_cast` COMPOUNDCURVE, MULTISURFACE or CURVEPOLYGON to GEOMETRYCOLLECTION (and back); #1573
 
 # version 0.9-7